implement launching StudentResponseFragment through notifications
StudentResponse has been renamed StudentResponseFragment to follow naming conventions more closely.

CallForStudentQuestionsResponseReceiver now has a static factory method getStartIntent(Context context) to generate the intent needed for the broadcast receiver to create the notification to launch the studentResponseFragment.

Notifications will now show up in the notification drawer when launched and will take the student to the UI for responding to a question when clicked.

Finally, a more modular way to launch certain fragments on the start of the MainActivity was added for convenience and readability.