make login fragment launch by default
The login fragment is now our default launch fragment to give the feel of a real, cohesive application. The debug fragment is still available for anyone who feels they need it. It can be accessed but clicking the Debug item in the setting menu (the three dots at the top right corner of the application)