Refactor broadcast receiver into a background service.
Using a service for this functionality makes more sense as a service will be able to remain alive in the background on the student client waiting to receive questions through the planned WIFI p2p interface. This commit also changes the min sdk back to 18 from 16. For some reason nelson changed this in a previous commit of his and I could not see where he noted a reason for the change, so I am changing it back because some of the methods I utilize in my fragment require API 18.