Add teacher service and corresponding socket actions
This commit adds the TeacherQuestionService to handle communication with Steven's StudentQuestionService and facilitate communication and transportation of the IQuestion object to the Student service and then recieve the answer back from the student's device. It implements the corresponding sockets mirroring Steven's functinality: TeacherSendQuestionAction and TeacherAcceptQuestionAction. These and the service implement the predefined interfaces established by both steven and cole as well as following the Template for a SocketAction.