Add answer collection from multiple devices
This commit adds multi-device answer collection.  Works simply by adding an answered question to the list of answered questions as they come in from students.  When the professor presses "collect", which is now renamed to "call time" answers are no longer accepted from student phones and and missing answers are considered unanswered and stored in the list following the new DefaultUnansweredQuestion format. There is also a getAnswerList() method which returns the list of answered questions for use by Thomas, can be removed later if we decide to only launch the results fragment from my fragment.