Add basic listview for visual reflection of answer collection
This commit adds a listview to the layout so that the answers can be shown on the TeacherQuestion fragment after "Call Time" is pressed.  Probably will be removed once everything is wired to Thomas.  Sets the listview as visible/invisible depending on whether the "Call Time" button has been pressed or if the "Send" button has been pressed.  Also updated the socket connection to check for an active outstanding question before accepting from student devices. For a quick fix also made the ArrayList of answered questions in TeacherQuestionService static for easy access for this demo, it will not stay this way!