Add basic wiring to TeacherQuestionResults
This commit should add basic functionality for sending the gathered answer list to Thomas' TeacherQuestionResults fragment after the call time button is pressed. I am unsure exactly of the interaction currently in place but just to get something up and running I think it will work. That being said there is much room for improvement and it can and will be cleaned up once the results are verified to work.