Clean up code and comment TeacherQuestion
This commit removes some redundant code and adds relevant comments to the TeacherQuestion fragment and service as well as removes the lingering visual elements of answers received on the TeacherQuestion fragment. It also puts in place a system for generating mock data so that if you send a question as a teacher and then press "Call Time" before any answers have been received 25 generic mult-choice questions are generated with hardcoded answer distributions.